Se crearon DTO necesarios para el mapeo de los objetos de Dominio, a la vez que se obtienen todos los valores del formato de Excel

+++ src/main/java/com/camunda/process/engine/dto/ControlCementoDTO.java
@@ -11,10 +11,10 @@ import java.util.Date;
 @Builder
 public class ControlCementoDTO {
 
-    private int saldo;
-    private int entradaKilos;
+    private double saldo;
+    private double entradaKilos;
     private Date fechaEntradaKilos;
-    private int salidaKilos;
+    private double salidaKilos;
     private Date fechaSalidaKilos;
     private Produccion produccion;
 }
